Umbrella clean-up/requirements.txt
requests
requests-oauthlib
oauthlib
python-dotenv
msal